+++ docs/ChangeLog.md
@@ -84,6 +84,7 @@
 - complete: requote for more compact representations on full completions `#D1700` 0000000
 - complete: improve support for `declare` and `[[ ... ]]` `#D1701` 0000000
   - syntax: fix completion and highlighting of `declare` with assignment arguments `#D1704` `#D1705` 0000000
+  - cmdspec: refactor `{mandb => cmdspec}_opts` `#D1706` `#D1707` 0000000
  
 ## Changes